Initial commit of (old) slurmpy code
slurmpy provides a one-to-one mapping of the Slurm C API
to Python with little higher-level abstraction mechanisms
(It tries to properly handle GC as much as possible).
This version of the code is already a bit older. It uses
an ad hoc parser utilizing regular expressions to extract
structures, typedefs, enums and functions. This only works
partially but can be used to create wrappers for a
significant subset of the Slurm API.