Use clangs AST to create the Python API
This commit reworks the parse-slurmh.py script. Instead
of an ad-hoc regexp-based mechanims we uses Clang's "tooling"
infrastructure to retrieve the required information from the
AST.

With this commit we also include slurmdb.h in the list
of headers that are considred.

The commit does not lead to more functionality in the Python
module. For this we will need to rewor write-pymodule.py which
will be done in a subsequent commit.